backend/requirements.txt
# ─── CyberHealth Scanner — Backend Dependencies ──────────────────────────────

# API Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
python-multipart==0.0.12   # requis par FastAPI pour les uploads de fichiers (Form, File)

# Rate Limiting
slowapi==0.1.9

# DNS Analysis
dnspython==2.7.0

# Async HTTP (pour futures intégrations : Shodan, VirusTotal…)
httpx==0.28.1

# Email (lead gen)
fastapi-mail==1.4.1

# Templating (rapports PDF)
jinja2==3.1.4

# PDF Generation (rapport expert)
weasyprint==62.3

# Logging structuré
structlog==24.4.0

# Variables d'environnement
python-dotenv==1.0.1

# Auth & Database
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# Google OAuth
google-auth==2.37.0
requests>=2.32.0     # requis par google.auth.transport.requests

# Paiement
stripe>=10.0.0

# Scheduler (monitoring hebdomadaire)
apscheduler==3.10.4

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
